requirements.txt
anyio==3.6.2
arrow==1.2.3
asttokens==2.2.1
autopep8==2.0.1
backcall==0.2.0
beautifulsoup4==4.11.2
blessed==1.20.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
chart-studio==1.1.0
click==8.1.3
comm==0.1.2
configparser==5.3.0
contourpy==1.0.7
cryptography==39.0.2
cycler==0.11.0
dash==2.6.2
dash-bootstrap-components==1.2.1
dash-core-components==2.0.0
dash-daq==0.1.7
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.6.6
decorator==5.1.1
diffusers==0.12.1
dominate==2.7.0
emoji==2.2.0
executing==1.2.0
fastapi==0.90.1
filelock==3.9.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-Compress==1.13
Flask-SeaSurf==1.1.1
Flask-SQLAlchemy==3.0.2
fonttools==4.38.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
inflect==6.0.2
inquirer==3.1.2
ipykernel==6.21.1
ipython==8.10.0
itsdangerous==2.1.2
jedi==0.18.2
jieba==0.42.1
Jinja2==3.1.2
jupyter_client==8.0.2
jupyter_core==5.2.0
kiwisolver==1.4.4
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
netifaces==0.11.0
numpy==1.24.2
OpenCC==1.1.6
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
plotly==5.13.0
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.4
Pygments==2.14.0
pymongo==3.11.0
pymssql==2.2.7
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
python-editor==1.0.4
pytz==2022.7.1
PyYAML==6.0
pyzmq==25.0.0
readchar==4.0.3
regex==2022.10.31
requests==2.28.1
retrying==1.3.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
sqlacodegen==3.0.0rc2
SQLAlchemy==1.4.46
stack-data==0.6.2
starlette==0.23.1
tenacity==8.2.1
tomli==2.0.1
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typing_extensions==4.4.0
ua-parser==0.16.1
unicorn==2.0.1.post1
urllib3==1.26.14
uvicorn==0.20.0
visitor==0.1.3
wcwidth==0.2.6
Werkzeug==2.2.2
zipp==3.13.0
flask_sqlalchemy==3.0.2
